Changes to be committed: 	modified:   SELFautorun.sh finally added some optargs stuff ^-^

+++ SELFautorun.sh
@@ -1,19 +1,66 @@
 #!/bin/bash
 
+
+#defaults
+batchloop=1
+batchsize=(10)
+events=1000
 config="/home/green642/sonic/CMSSW_12_5_0_pre4/src/HeterogeneousCore/SonicTriton/data/models/particlenet_AK4_PT/config.pbtxt"  # Replace with your file name
+#defaults
 
-batchsize=("10" "25" "50")  # Replace with your desired numbers
+help(){
+    echo "autoRun [options]"
+	echo 
+	echo "Options:"
+	echo "-b          Choose batch sizes to run (Default: (10)) (Input as a list)"
+    echo "-e          Choose how many events per inference (Default: 1000)"
+    echo "-h          Print this message and exit"
+	#echo "-d [dir]    Directory containing run.py"
+	echo "-r          Number of runs per batch size"
+   # echo "-o [name]   Choose output file for time report data (default: output.txt)"
+   # echo "-s"         Show output
+	exit $1
+}
+while getopts "b:d:e:hr:o:s" opt; do
+case "$opt" in
+b)
+batchsize="$OPTARG"
+;;
+e)
+events="$OPTARG"
+;;
+h)
+help 0
+;;
+r)
+batchloop="$OPTARG"
+;;
+esac
+done
 
-rm tempoutput.txt
-rm output.txt
+if [[-f output.txt]]; then
+    rm output.txt
+fi
+
 
 cd  /home/green642/sonic/CMSSW_12_5_0_pre4/src/sonic-workflows
+source /cvmfs/cms.cern.ch/cmsset_default.sh
+cmsenv
 for number in ${batchsize[@]}; do
     echo 'Starting with '${number}''
-    sed -i "6 s/\[ [0-9]* \]/[ $number ]/g" "$config" 
+    sed -i "6 s/\[ [0-9]* \]/[ $number ]/g" "$config" #changes the config line from [x] to [batchsize]
+   #sed -i (edit the current config file instead of making a copy)
+   #"line 6,  substitute/ [ 1 or more 0-9 digits ]/number/global"
+   #global as in, do it for everything on the line. might not need this.
     echo -e 'Preferred Batch size: '${number}' \n' >> output.txt
-    cmsRun run.py maxEvents=1000 threads=4 device=gpu tmi=True  &> tempoutput.txt | sed -n '/TimeReport>/,/^Mem/ { /^Key/ d; p; }' >> output.txt
-   
+    for i in {1..$batchloop}; do
+        cmsRun run.py maxEvents=${events} threads=4 device=gpu tmi=True 2>&1| tee tempoutput.txt #the 2>&1 redirects stderr (which has TimeReport) to stdout
+        sed -n '/TimeReport>/,/^Mem/ { /^Key/ d; p; }' tempoutput.txt >> output.txt
+        echo "loop '$i' of '$batchloop'"
+    done
     echo -e ' \n\n\n' >> output.txt
     echo 'Done with '${number}''
+
+    rm tempoutput.txt
+
 done
